move overview view to phenopackets app
